[refactor] Migrate LSN Type from usize to u16 and Optimize SSA Logging
This commit introduces a comprehensive migration of LSN (Logical Sequence Number) from usize to u16 across the SSA (Static Single Assignment) infrastructure:

- Replace usize with u16 in SSALogger, ShadowStack, and related types
- Update graph and dependency tracking to use u16 for LSN representation
- Modify input and output handling to work with u16 LSNs
- Optimize memory usage by reducing LSN type size
- Adjust input resolution and dependency tracking methods
- Update logging and tracing components to support u16 LSNs

The changes aim to reduce memory overhead and improve performance in SSA logging and graph construction.

+++ crates/revm-ssa-graph/src/tracer.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 /// Mismatch record
 #[derive(Debug, Clone)]
 pub struct MismatchRecord {
-    pub lsn: usize,
+    pub lsn: u16,
     pub opcode: u8,
     pub original: Vec<SSAOutput>,
     pub graph: Vec<SSAOutput>,
@@ -14,7 +14,7 @@ pub struct MismatchRecord {
 #[derive(Debug)]
 pub struct ExecutionTracer {
     /// Original execution results
-    original_results: HashMap<usize, Vec<SSAOutput>>,
+    original_results: HashMap<u16, Vec<SSAOutput>>,
     /// Mismatch records
     mismatches: Vec<MismatchRecord>,
 }
@@ -28,12 +28,12 @@ impl ExecutionTracer {
     }
 
     /// Record original execution result
-    pub fn record_original(&mut self, lsn: usize, outputs: Vec<SSAOutput>) {
+    pub fn record_original(&mut self, lsn: u16, outputs: Vec<SSAOutput>) {
         self.original_results.insert(lsn, outputs);
     }
 
     /// Record graph execution result
-    pub fn record_graph(&mut self, lsn: usize, outputs: Vec<SSAOutput>, opcode: u8) {
+    pub fn record_graph(&mut self, lsn: u16, outputs: Vec<SSAOutput>, opcode: u8) {
         // Check if results match
         if let Some(original) = self.original_results.get(&lsn) {
             if !Self::compare_results(original, &outputs) {
